prevent admin from being removed as admin

+++ src/routes/members/{user_id}/roles/{role_id}/delete.ts
@@ -47,6 +47,11 @@ const deleteController = async (req: Request, res: Response): Promise<void> => {
         return;
     }
 
+    if (ur.prevent_delete) {
+        res.sendStatus(405);
+        return;
+    }
+
     await dbConnection.manager.delete(UserRole, {
         user_id: Number(id),
         role_id: Number(role_id),
